backend/mind_body_connection/requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
joblib==1.4.2
nltk==3.9.1
numpy==2.1.3
packaging==24.2
pandas==2.2.3
patsy==1.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.11.6
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sqlparse==0.5.2
statsmodels==0.14.4
textblob==0.18.0.post0
threadpoolctl==3.5.0
tqdm==4.67.1
tzdata==2024.2
